add profile back-end structure and FREE and PREMIUM users

+++ api/src/main/java/com/fabio/flashcards_app/data/dto/deck/DeckResponseDTO.java
@@ -8,7 +8,7 @@ public record DeckResponseDTO(
         String subject,
         Boolean isPublic,
         Boolean reviewEnabled,
-        Integer cardCount,
+        Long cardCount,
         Long folderId
 ) {}
